requirements.txt
streamlit==1.31.1
Pillow
pdf2image
google
